requirements-froze.txt
google-api-python-client==2.110.0
python-dotenv==1.0.0
yt-dlp==2025.7.21
openai-whisper==20250625
openai==1.97.1
numpy==2.2.6
scikit-learn==1.7.1
ffmpeg-python==0.2.0
sentence-transformers==5.0.0
torch==2.7.1
transformers==4.53.3
bertopic==0.17.3
ruptures==1.1.9
umap-learn==0.5.9.post2
hdbscan==0.8.40
plotly==6.2.0
networkx==3.5
SQLAlchemy==2.0.41
psycopg2-binary==2.9.10
